requirements.txt
Flask
Flask-Cors
google-generativeai
pandas
sentence-transformers
faiss-cpu
numpy